requirements.txt
requests
lxml
pandas
discord.py
asyncio